Initial attempt at a game layout
It would be nice to take advantage of the extra row on the Atreus,
compared to the Vortex Core, to create a number row. Additionally,
some people say ESDF gaming works better than WASD on a keyboard like this.

This takes a QWERTY layout and shifts it right one column and down one row.
Space and tab stay in the same place and v is moved to where tab would usually be.